pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "cc-fetcher"
version = "0.1.0"
description = "Fetches raw data from various APIs and web sources for the China Compass pipeline"
authors = ["China Compass Team"]
readme = "README.md"
packages = [{include = "fetcher", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27"
feedparser = "^6.0"
beautifulsoup4 = "^4.12"
yfinance = "^0.2"
pyyaml = "^6.0"
click = "^8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"
ruff = "^0.5"
respx = "^0.21"

[tool.poetry.scripts]
fetcher = "fetcher.cli:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
